refactor(ms2): removed folder creation button
